requirements.txt
langchain 
langchain-community 
langchainhub 
langchain-openai 
chromadb 
bs4
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib      
pypdf
fitz
pandas
PyMuPDF
pdfplumber 
requests  
pyarrow
matplotlib
plotly
scipy
scikit-learn
tenacity
openai
numpy
Flask
